+++ lib/src/dart/extensions.dart
@@ -7,5 +7,3 @@ export 'num_extension.dart';
 export 'int_extension.dart';
 
 export 'string_extension.dart';
-
-export 'object_extension.dart';