+++ lib/src/dart/int_extension.dart
@@ -3,7 +3,7 @@ import 'extensions.dart';
 ///Int extension to extend List functionality
 extension IntExtension on int {
   /// Get Month name
-  String getMonthName([bool isShort = false]) => match(<int, String>{
+  String getMonthName([bool isShort = false]) => <int, String>{
         1: isShort ? 'Jan' : 'January',
         2: isShort ? 'Feb' : 'February',
         3: isShort ? 'Mar' : 'March',
@@ -16,10 +16,10 @@ extension IntExtension on int {
         10: isShort ? 'Oct' : 'October',
         11: isShort ? 'Nov' : 'November',
         12: isShort ? 'Dec' : 'December',
-      }, 'Invalid day of month');
+      }.match(this, 'Invalid day of month');
 
   /// Get Week name
-  String getWeekName([bool isShort = false]) => match(<int, String>{
+  String getWeekName([bool isShort = false]) => <int, String>{
         1: isShort ? 'Mon' : 'Monday',
         2: isShort ? 'tue' : 'Tuesday',
         3: isShort ? 'Wed' : 'Wednesday',
@@ -27,5 +27,5 @@ extension IntExtension on int {
         5: isShort ? 'Fri' : 'Friday',
         6: isShort ? 'Sat' : 'Saturday',
         7: isShort ? 'Sun' : 'Sunday',
-      }, 'Invalid day of week');
+      }.match(this, 'Invalid day of week');
 }